refactor(migrations): moved onInteraction, onHover , onViewport to core/primitives
